@@ -82,6 +82,6 @@ public class ContributingGuidelinesProbe extends Probe {
 
     @Override
     public long getVersion() {
-        return 1;
+        return 2;
     }
 }
